pyproject.toml
[project]
name = "bilibili-to-text"
version = "0.1.0"
description = "Bilibili 视频转文字工具：下载音频 → 转录 → Markdown → LLM 总结"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alibabacloud-oss-v2>=1.2.4",
    "dashscope>=1.25.12",
    "fastapi>=0.116.1",
    "groq>=1.0.0",
    "litellm>=1.74.9",
    "minio>=7.2.12",
    "openai>=2.17.0",
    "pydub>=0.25.1",
    "questionary>=2.0.0",
    "rich>=13.7.1",
    "reportlab>=4.4.9",
    "requests>=2.31.0",
    "textual>=0.45.1",
    "uvicorn>=0.35.0",
    "yutto>=2.1.1",
    "pytest-playwright>=0.7.2",
]

[tool.setuptools.packages.find]
include = ["b2t*"]

[project.scripts]
b2t = "b2t.cli:main"
md-table-pdf = "b2t.converter.md_table_to_pdf:main"
